Removed utils subdirectory in tests
